% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_completed}
\alias{get_completed}
\alias{get_completed.rdmc_tuned}
\alias{get_completed.soft_impute_tuned}
\alias{get_completed.median_impute}
\alias{get_completed.mode_impute}
\alias{get_imputed}
\title{Extract the completed (imputed) data matrix}
\usage{
get_completed(object, ...)

\method{get_completed}{rdmc_tuned}(object, ...)

\method{get_completed}{soft_impute_tuned}(object, discretized = FALSE, ...)

\method{get_completed}{median_impute}(object, discretized = FALSE, ...)

\method{get_completed}{mode_impute}(object, ...)

get_imputed(object, ...)
}
\arguments{
\item{object}{an object returned by a matrix completion algorithm.}

\item{\dots}{additional arguments to be passed down to methods.}

\item{discretized}{a logical indicating if the imputed data matrix with or
without the discretization step should be extracted.}
}
\value{
The completed (i.e., imputed) data matrix.
}
\description{
Extract the completed (i.e., imputed) data matrix from an object returned by
a matrix completion algorithm.
}
\note{
Matrix completion and imputation are synonymous terms used in
different streams of the literature, hence \code{get_imputed()} is an
alias for \code{get_completed()} with the same functionality.
}
\seealso{
\code{\link{rdmc_tune}()}, \code{\link{soft_impute_tune}()},
\code{\link{median_impute}()}, \code{\link{mode_impute}()}
}
\author{
Andreas Alfons
}
\keyword{utilities}
