% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_nb_iter}
\alias{get_nb_iter}
\alias{get_nb_iter.rdmc_tuned}
\alias{get_nb_iter.rdmc}
\title{Extract the number of iterations}
\usage{
get_nb_iter(object, ...)

\method{get_nb_iter}{rdmc_tuned}(object, ...)

\method{get_nb_iter}{rdmc}(object, ...)
}
\arguments{
\item{object}{an object returned by an iterative matrix completion
algorithm.}

\item{\dots}{currently ignored.}
}
\value{
The number of iterations performed in the iterative algorithm.
}
\description{
Extract the number of iterations from an object returned by a matrix
completion algorithm.
}
\seealso{
\code{\link{rdmc_tune}()}
}
\author{
Andreas Alfons
}
\keyword{utilities}
