% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_X}
\alias{get_X}
\alias{get_X.rdmc_tuned}
\alias{get_X.soft_impute_tuned}
\alias{get_X.median_impute}
\alias{get_X.mode_impute}
\title{Extract the imputed data matrix}
\usage{
get_X(object, ...)

\method{get_X}{rdmc_tuned}(object, ...)

\method{get_X}{soft_impute_tuned}(object, discretized = FALSE, ...)

\method{get_X}{median_impute}(object, discretized = FALSE, ...)

\method{get_X}{mode_impute}(object, ...)
}
\arguments{
\item{object}{an object returned by a matrix completion algorithm.}

\item{\dots}{additional arguments to be passed down to methods.}

\item{discretized}{a logical indicating if the imputed data matrix with or
without the discretization step should be extracted.}
}
\value{
The imputed data matrix.
}
\description{
Extract the imputed data matrix from an object returned by a matrix
completion algorithm.
}
\seealso{
\code{\link{rdmc_tune}()}, \code{\link{soft_impute_tune}()},
\code{\link{median_impute}()}, \code{\link{mode_impute}()}
}
\author{
Andreas Alfons
}
\keyword{utilities}
