% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmc-deprecated.R
\name{rdmc-deprecated}
\alias{rdmc-deprecated}
\alias{get_X}
\title{Deprecated functions in package \pkg{rdmc}}
\usage{
get_X(object, ...)
}
\description{
The following functions are deprecated and may be removed as soon as the
next version.
}
\section{Replacement functions}{

Instead of \code{get_X()}, use \code{\link{get_completed}()} or
\code{\link{get_imputed}()}.
}

\keyword{internal}
