% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmc-package.R
\docType{package}
\name{rdmc-package}
\alias{rdmc-package}
\title{\packageTitle{rdmc}}
\description{
\packageDescription{rdmc}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{rdmc}
\packageIndices{rdmc}
}
\references{
Archimbaud, A., Alfons, A., and Wilms, I. (2024) Robust Matrix Completion
for Discrete Rating-Scale Data. arXiv:2412.20802.
\doi{10.48550/arXiv.2412.20802}.

Hastie, T., Mazumder, R., Lee, J. D. and Zadeh, R. (2015) Matrix Completion
and Low-Rank SVD via Fast Alternating Least Squares. \emph{Journal of
Machine Learning Research}, \bold{16}(104), 3367--3402.

Mazumder, R., Hastie, T. and Tibshirani, R. (2010) Spectral Regularization
Algorithms for Learning Large Incomplete Matrices. \emph{Journal of Machine
Learning Research}, \bold{11}(80), 2287--2322.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aalfons/rdmc}
  \item Report bugs at \url{https://github.com/aalfons/rdmc/issues}
}

}
\author{
\packageAuthor{rdmc}

Maintainer: \packageMaintainer{rdmc}
}
\keyword{package}
