% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_impute.R
\name{mode_impute}
\alias{mode_impute}
\title{Mode imputation}
\usage{
mode_impute(X, values = NULL)
}
\arguments{
\item{X}{a matrix or data frame of discrete data with missing values.}

\item{values}{an optional numeric vector giving the possible values.
Currently, the possible values are assumed to be the same for all columns.
If \code{NULL}, the unique values of the observed parts of \code{X} are
used.}
}
\value{
An object of class \code{"mode_impute"}.

The class structure is still experimental and may change.  Use the accessor
function \code{\link{get_X}()} to extract the imputed data matrix.
}
\description{
Perform mode imputation for discrete data.  In case of multiple modes in a
given column, one of them is selected at random for each missing cell.
}
\note{
The mode is computed as the most frequent value, hence this function is only
suitable for discrete data.  It does not estimate the mode of a continuous
density.
}
\seealso{
\code{\link{median_impute}()}
}
\author{
Andreas Alfons
}
\keyword{multivariate}
