% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmc-package.R
\docType{package}
\name{rdmc-package}
\alias{rdmc-package}
\title{\packageTitle{rdmc}}
\description{
\packageDescription{rdmc}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{rdmc}
\packageIndices{rdmc}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aalfons/rdmc}
  \item Report bugs at \url{https://github.com/aalfons/rdmc/issues}
}

}
\author{
\packageAuthor{rdmc}

Maintainer: \packageMaintainer{rdmc}
}
\keyword{package}
